// frontend/src/app/shared/generated-api/model/checklistItemInput.ts
/**
 * Glacier Notes Cloud API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ChecklistItemInput { 
    id?: string;
    text: string;
    checked: boolean;
}